genmesh: [Feature]: Implement Periodic Inlet and Outlets.
The idea is to simulate the inlet, "n" sections of the periodic column,
and then the outlet.

In order to implement this I had to refactor a lot of the code into
separate classes. So now we have (broadly):

Geometry: Handles Geometry creation and operation
Column: Takes in dimtags of fragmented column geometry and
    - takes care of surfaces and volumes
    - adding physical names and groups
    - writing out
    - periodicity

The Model class, which was previously a "God" class, is now just a
driver.

Some features are broken with this commit because I was too lazy to fix
them along the way: 1) differring fragment format and main format and 2) geometry input.

Generating gradient beads works for lc_beads < lc_out, but not the other
way around.

_surfaces_beads probably still contain the top and bottom cut beads
surfaces.

There are definitely a lot of memory leaks that need fixing.

I'll get to fixing these issues in future commits.